refactor: move initialise solvers to benchmark_utils.py #927
