+++ benchmark/blobs_benchmark.py
@@ -111,6 +111,7 @@ def setup_solvers(
 
     :param coreset_size: The size of the coresets to be generated by the solvers.
     :param sq_exp_kernel: A Squared Exponential kernel for KernelHerding and RPCholesky.
+        The squared root kernel for KernelThinning is also derived from this kernel.
     :param stein_kernel: A Stein kernel object used for the SteinThinning solver.
     :param delta: The delta parameter for KernelThinning solver.
     :param random_seed: An integer seed for the random number generator.
